backend/form_builder_service/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend API"
authors = ["Mustafa Abdallah <mustafa.abdulluh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
python-decouple = "^3.8"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
django-storages = "^1.13.2"
google-cloud-storage = "^2.8.0"
google-auth = "^2.17.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
